Add pylint, mypy and pycodestyle (former PEP8) tests to CircleCI.
We use a generated .pylintrc. Important notices:
* Linting members of numpy and twisted.internet.reactor is disabled due
to a bug in pylint. See https://github.com/PyCQA/pylint/issues/779
* We ignore minilight since it's a forked third-party codebase
* For the configuration ease we enabled `unsafe-load-any-extension`.
This makes it trivial to lint C modules. Alternatively, we could simply
whitelist them with `extension-pkg-whitelist` but this would be tedious.
* Old findings will not trigger a CI fail. Please note that it's a
fairly diffucult topic to decide which finding is old and which new, so
we assume the diff output to be the jury.
* The list of disabled checks is intented to be small. We will disable
the check we find useless/inappropriate/overzealous as we encounter
them in the code.

Additionally, a malformed type annotation was fixed in
`apps/rendering/resources/imgrepr.py`.

The `setup` subdirectory has been renamed into `setup_util` since it
conflicts with `setup.py` and confuses the linters.