Use different directories for CompTaskKeeper and TaskManager persistence.
With previous implementation TaskManager and CompTaskKeeper used the
same directory for pickling their state. As a result, TaskManager tried
to read the CompTaskKeeper's state, failing due to a type mismatch.

The persistence is still disabled by default. Interfaces should expose
an option to enable or disable it at the user's discretion.