Unsupported task stats #1448.
A new command "tasks unsupport" has been added to client that shows
statistics for unsupported tasks that the node currently sees in
the network. When this command is followed with a number of days,
the statistics are shown for all tasks seen by this node during
the last number of calendar days. The statistics are persistent.

Change-Id: Ia4a32e4dbe598eccb16b48bb73db2d7a0cefb318