Task persistance a.k.a. Code for saving task manager state (#1581)
* A little of refactor in task manager and task keeper

* Annotate types in taskmanager.py

* Remove the unused, monkey-patched field `task_status` from `Task`s.

In `TaskManager`, there was a monkey-patched field `task_status` used,
essentially a duplicate of the `status` field in the `TaskState` class.

Since it was not used anywhere, it was removed from the codebase.

Additionally, the tests were refactor to use the `unittest` methods
for assertions instead of the plain `assert`.

* Use different directories for CompTaskKeeper and TaskManager persistence.

With previous implementation TaskManager and CompTaskKeeper used the
same directory for pickling their state. As a result, TaskManager tried
to read the CompTaskKeeper's state, failing due to a type mismatch.

The persistence is still disabled by default. Interfaces should expose
an option to enable or disable it at the user's discretion.

* Continuation of marmistrz refactoring

* Clean typechecks

* Clear mocks

* Clean merge

* wip

Signed-off-by: ggruszczynski <ggruszczynski@golem.network>

* Remove old tasks in CompTaskKeeper

* Remove mock dependency

* PEP8

* Remove unused ttl references

* Cleanup

* Clean tasks_dir var name

* Clean up of unrelated changes

* Clean up of unrelated changes

* Remove dump on task delete

* added dump_and_restor_unit_test

Signed-off-by: ggruszczynski <ggruszczynski@golem.network>

* added test for compt task keeper -> remove old tasks

Signed-off-by: ggruszczynski <ggruszczynski@golem.network>

* (update) added test for compt task keeper -> remove old tasks

Signed-off-by: ggruszczynski <ggruszczynski@golem.network>

* moved dispatcher.send outside with and try-except block

Signed-off-by: ggruszczynski <ggruszczynski@golem.network>

* Log dumping task_id

* Remove task_status changes