Collecting task requestor stats, #1492
Statistics about tasks requested by the node are gathered by
RequestorTaskStatsManager objects. They are collected from
golem.taskmanager signals with task_status_updated event, modified in
the previous commit.

Note that statistics are gathered per session and each restart of the
client will introduce some errors. This is not considered to be a
problem for their intended use, that is diagnosing persistent problems
with clients.

This is not hooked into the rest of Golem yet, i.e. there are no
instances of RequestorTaskStatsManager created. It will be done in the
next commit.

Change-Id: Ife46ef674f310ab983618af3aa4e2067fc0212ee