Synchronization of node time with monitor (#2038)
Local time of node is included in ping-me request. Monitor sends back
difference between received timestamp and its own time. Of course this
is **not** the exact difference between node's time and monitor's time
because it includes network latency. If the received value exceeds
defined maximum (10 seconds for now) a warning is displayed and an
`unsynchronized` event is published to `golem.p2p` channel.

The whole ping-me feature was significantly refactored:
  * `listening` event has now `ports` argument instead of `port` (it was
    already representing multiple ports so the name was misleading)
  * requests sent to `/ping-me` endpoint have `ports` field instead of
    `port` as well
  * response from `/ping-me` is expected to contain `port_statuses`: an
    array of objects of the following format:
    ```
    {
      "port": 12345,            // port number
      "is_open": false,         // is the port ping-able
      "description": "timeout"  // description (e.g. "open", "timeout")
    }
    ```
  * `Node` contains a nice `port_statuses` dict instead of messy
    `port_status` all-containing string

Tests for ping-me were improved.

Because of API changes `PROTO_VERSION` was bumped to 1.