Golem terms of use (#2345)
User has to accept terms of use before running Golem. If they don't the
client will not connect to the network and an error message will be
displayed in the logs. Terms of use can be displayed and accepted via
CLI. There are RPC methods to check if terms have been accepted and to
accept them. These methods could be used by the GUI to provide a more
user-firendly way of accepting the terms.

Information about acceptance is stored in the databse.

There is no need to restart Golem after accepting the terms.