Golem terms of use (#2373)
User has to accept terms of use before running Golem. If they don't the
client will not connect to the network and an error message will be
displayed in the logs. Terms of use can be displayed and accepted via
CLI. There are RPC methods to display the terms, to check if terms have
been accepted, and to accept them. These methods could be used by
the GUI to provide a more user-firendly way of accepting the terms.
Information about acceptance is stored in the database. There is no need
to restart Golem after accepting the terms. Terms are versioned so that
after updating them user could be prompted for re-acceptance.