PoC p.12 Fix taskserver tests in CI
DockerEnvironment checks if it's docker images are present before
it can be selected for a task.

This causes taskserver tests to fail on CI, because they are using
DummyTaskEnvironment (which is a subclass of DockerEnvironment) and
"Dummy" task type in order to perform checks of task_server. On CI
there is no docker support and as a result tests that were using "Dummy"
task type were failing on CI.

This commit mocks _check_software in DummyTaskEnvironment so that taskserver
tests continue to work on CI.

Change-Id: Ief827403c3107710940f6dbfaf243b244fc4d667