PoC firejail (#2848)
* PoC: Added a method to extract paths from docker images.

Change-Id: I70870441b8d965a0c1db13ec1be7a7b5ea88d158

* PoC p.1 Tasks no longer specify env, Provider selects env from task_type

Up til now specification how task will be executed was included in task itself
by specifying docker_images and environment in task header (and some minor data in CTDs).
Provider had no saying about how to calculate a given task.

This commit moves the responsibility of deciding how to calculate a task from requestor
to provider. Tasks will now only hold data required for the computation (parameters, requestor
data, task type). Provider based on task type and internal config can select appropriate environment
for computation. This allows different providers to use different computation methods for the same
task.

This task is the initial refactoring for rendering blender tasks outside of docker.

Change-Id: I74a9a521c85ca61b5b5bce0cbb162a1935b68b41

* PoC p.2 Firejail actually runs blender

In this commit new Firejail environment and TaskThread is created for
running blender tasks. In order to use it user needs to manually
specify correct FirejailEnvironment in registered_envs.ini.

Since blender now supports two different strategies for computation
I've also moved BlenderEnvironment-related files to point out
what is generic for a blender task and what is specific for each
computation strategy.

Change-Id: I85faf4f909ea021d01beacdffb66629961912a5d

* PoC p.3 Blender through firejail tries to run using GPU

"Tries" because although render script is told to turn on GPU rendering (CUDA or OPENCL),
it is still possible that local blender does not turn it on and will render on CPU.

Change-Id: Icbf6e61b4b8b0d22ee5ccb9690f1aa5f2126152f

* PoC p.4 Extract blender from docker image

That way we can guarantee the correct version of blender running evaluation

Change-Id: Ief426f64d780704eb52da34d518c9b26236ce212

* PoC p.5 Pre-run blender to make sure kernel modules are loaded

Running blender through firejail prohibits loading of kernel modules. This is
unfortunate as initializing CUDA makes nvidia drivers load kernel modules
on demand. This change ensures blender is run once without firejail.

Blender is run at the time of extraction from docker image. Since it is
extracted to temporary directory of the system, is should be run at least
once after system reboot.

There are no protections agains user removing the modules in a running system
- they will not be reinitialized.

Change-Id: Id86d47c3db2cc78778f7db77afde72db6c7c3137

* PoC p.6 Requirements and Supports

This commit introduces a mechanism by which tasks can request
specific capabilities in computing environment and at the same
time environments can now claim to support specific requirements.

From the task's side a new class of Requirement is added, task can specify
any number of requirements - those will be sent in TaskHeader allowing
all nodes to take requirements into account when requesting new subtasks.

From the environment's side a Support class is added, environment can
have any number of Supports.

At the moment of looking for a matching environment to calculate a task
not only task_type is checked, but also if all of task's requirements
are met by environment's supports.

Change-Id: I549ef50f77dcdc272573df6c50d8c364fa60084c

* PoC p.7 Run firejail with explicit profile

Change-Id: Ic5f42c73bb2bbe012ff0cdaa927d1bdcae1fb52e

* PoC p.8 Pass cpu and memory limits to firejail task thread

Change-Id: Ia4dfe15616b4ff1548c601efb926ea11097e65b8

* PoC p.9 Add firejail test

Change-Id: I3a479d3c6a84bf7305401ef37d46b30f78baab23

* PoC p.10 Fix for "Too many open descriptors"

On MacOS tests were failing due to "too many open file descriptors".

Turns out the problem was with the way blender benchmarks used to get temporary
file for benchmark's output. The method they were using was leaving an opened os-level
handle to the temporary file. Problem was also present in production version of golem,
it was not only limited to tests. Previously golem's tests by happy accident kept the
number of opened file descriptors below the limit, with our changes to benchmarks
the problem just got more visible.
 
Change-Id: I988a660c275d933d14d9b4570cd28840d21c8ab2

* PoC p.11 Fix linting issues

Change-Id: I714d0ef383789cffb450a6de60bf046fa60e3b5d

* PoC p.12 Fix taskserver tests in CI

DockerEnvironment checks if it's docker images are present before
it can be selected for a task.

This causes taskserver tests to fail on CI, because they are using
DummyTaskEnvironment (which is a subclass of DockerEnvironment) and
"Dummy" task type in order to perform checks of task_server. On CI
there is no docker support and as a result tests that were using "Dummy"
task type were failing on CI.

This commit mocks _check_software in DummyTaskEnvironment so that taskserver
tests continue to work on CI.

Change-Id: Ief827403c3107710940f6dbfaf243b244fc4d667