Sharing performance information
Estimated performance (bechmark results) is shared in the `metadata`
field of `Hello` message. In order to achieve this 'metadata providers'
were added to P2PService.

Received metadata from other nodes are stored in the `KnownHosts` table
in the database. Performance information obtained this way are used to
computed the potential number of providers for netmasking (instead of
relying on hard-coded list).