Wait for benchmarks during client initialization
Since benchmarks are running in a separate thread it was possible that
node started talking to others before the bechmarks were complete. This
would result in missing bechmark scores being sent in `Hello.metadata`
which would not be updated later because `Hello` message is only sent at
the beginning of session.