Start VM means `start` not `restart`
Changed the handling of `start` command in DockerMachineCommandHandler
so that it calls `docker-machine start` instead of `restart`. It was
not working correctly because Hyper-V driver for docker-machine tries
to connect to a VM via SSH before stopping it and this would hang up
forever if the VM is not running in the first place.

Added a separate `restart` command which is used in `_recover()`.