Added warning on cli and app net mismatch
This adds a check in golemcli which attempts to detect whether an
instance of golem app is running on the system with a different target
chain (e.g. CLI is started with --mainnet flag, but there's a golem
instance running on testnet). To alert the user of a potential mismatch,
a warning message is printed.