Added warning on cli and app net mismatch (#3540)
This adds a check in golemcli to see if there's an instance of golem core running on the system which uses a different chain compared to CLI (e.g. CLI is configured for mainnet and core uses testnet). In case of a potential mismatch, a warning message is logged.
Checking for a running golem core is based on the appropriate datadir being locked. To enable this, the logic for acquiring a lock on datadir was refactored and moved to golemapp (so that datadir is locked as early as possible, before it's being actually used by the app).