Check if there's enough disk space to dump VM memory
Setting "save" as host shutdown action for VM requires pre-allocating
disk space for memory dump. The VM will not start if there is not enough
space on system drive. Therefore we check the available disk space
during VM reconfiguration and fall back to "shutdown" action if there is
not enough.