Hyper-V: Check disk space before VM creation
Docker-machine sets host shutdown action to 'save' and there is no
option to override it. This is the preferred action but using it
requires pre-allocating disk space for memory dump on VM startup. If
there is not enough space we should try to create the VM with minimal
amount of RAM and reconfigure it later to use 'shutdown' host shutdown
action.