Use native SSH in docker-machine
SSH connections to VMs were very slow due to bug in OpenSSH. Actually,
every time ssh was called it would wait for the amount of time specified
by the ConnectTimeout option (which obviously **is not** the expected
behaviour for this option).

Because the ConnectTimeout was hard-coded in docker-machine to 10
seconds we could not reconfigure it. Instead we decided to use the
--native-ssh option for docker-machine which doesn't use OpenSSH at all
but the native (Go-based) implementation.

Since this option significantly speeded up starting VM we don't need the
custom script for doing it.