Interfaces for environments refactoring
Used `golem.envs` instead of `golem.environments` because we don't want
to replace the old environments *yet*. The old package will be removed
when we are done.

Environment events should probably be somehow defined.

We can think of using `NewType` for all ID types so that one would have
to use the type explicitly.