Fixed Hyper-V VM remaining in invalid state
If creating the machine failed during virtual disk creation then the VM
would remain in a an invalid state (actually not the VM but some config
files used by docker-machine). This in turn would make it impossible to
re-create the VM.
The error was unsuccessfully mitigated against by adding a remove
command in HyperVHypervisor._failed_to_create() method. This was not
effective because removing the VM sometimes fails if it done shortly
after the unsuccessful creation.
To make the cleanup more robust a retry loop was added. It required
changing the logic of DockerMachineHypervisor.vms() because `docker-
machine remove -q` doesn't print out VMs in invalid states.