Models for new environment performance
As we will be introducing new environments we need to adjust the
database schema to store performance scores for them. Performance model
cannot be re-used because a clean separation between old and new
environments is needed during transition period. For the same reason
performance field was introduced to the KnownHosts model.