Don't defer docker whitelist check to another thread
This check is not time-consuming so it can be performed in the calling
thread instead of being deferred to another thread.

Actually, this is a workaround for integration tests failing on Windows.
For some reason if database was used in another thread (as it happened
before this change) the unit test cleanup would fail because "The
golem.db is used by another process and cannot be removed".