New logic for setting up and cleaning up environments
EnvironmentManager now ensures that exactly one environment is active at
a time. Once an environment is needed it is automatically prepared. It
would be cleaned up when another environment is going to be used.

This brings a disadvantage of keeping up idle environments but allows to
avoid lengthy setups and cleanups if one environment is used a few times
consecutively.

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>