Add deadlines to message queue entries (#5106)
Adds a deadline field to the the database model QueuedMessage. Extends
message queue logic to discard messages which are past their deadline.

When putting new messages to the queue an optional timeout parameter
can be specified. A message's deadline is derived from this timeout.
This change also adds such a timeout to ResourceHandshakeStart and
WantToComputeTask messages.